requirements.txt
pyodbc
python-dotenv
azure-eventhub
azure-functions
azure-storage-blob